adds diagonal movement to the game
I changed the character class inorder to accomodate diagonal motion. It
originally reset the entire velocity vector everytime a key press was
input but now it only updates the projection of the velocity vector onto
the unitvector of whatever arrow key is pressed. This commit also added
in a new library to the project in order to get access to the entire
keyboard state array which was nessicary to get the states of more than
one key at a time.